requirements.txt
flask
flask-mail
nltk
scikit-learn
PyPDF2
flask-cors
python-docx